+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Fixed an issue where material parameters were not annotated correctly in decompiled shaders.
 * Fixed an issue where material parameters were not handled correctly in the shader JSON.
 * Fixed an issue where some Xenoblade 3 models used incorrect vertex skinning weights.
+* Fixed an issue where Xenoblade 1 DE and Xenoblade 2 models did not load the high resolution base mip level.
 
 ## 0.2.0 - 2023-11-22
 ### Added